requirements.txt
asgiref==3.7.2
certifi==2023.11.17
charset-normalizer==3.3.2
Django==4.2.8
django-rest-framework==0.1.0
djangorestframework==3.14.0
dnspython==2.5.0
idna==3.6
pymongo==4.6.1
python-decouple==3.8
pytz==2023.3.post1
requests==2.31.0
sqlparse==0.4.4
typing_extensions==4.9.0
urllib3==2.1.0
